+++ apps/web/src/components/section-dashboard.tsx
@@ -28,10 +28,10 @@ export default function DashboardSection({
   children,
 }: SectionProps) {
   return (
-    <Card className="h-full w-full gap-2 border-none bg-transparent p-2 group-has-data-[collapsible=icon]/sidebar-wrapper:gap-0 group-has-data-[collapsible=icon]/sidebar-wrapper:p-0">
+    <Card className="relative z-10 h-full w-full gap-2 rounded-none border-none bg-transparent p-2 group-has-data-[collapsible=icon]/sidebar-wrapper:gap-0 group-has-data-[collapsible=icon]/sidebar-wrapper:p-0">
       {title && description && (
         <CardHeader className="p-0">
-          <div className="flex flex-row">
+          <div className="flex flex-row items-center">
             <GoBackButton />
             <div className="ml-4 h-full w-full">
               {title && <CardTitle>{slugToTitle(title)}</CardTitle>}
@@ -51,8 +51,8 @@ export default function DashboardSection({
           </CardAction>
         </CardHeader>
       )}
-      <div className="h-[calc(100vh-11rem)] overflow-auto">
-        <CardContent className="p-0">{children}</CardContent>
+      <div className="h-[calc(100dvh-12rem)] overflow-auto">
+        <CardContent className="px-0 py-2">{children}</CardContent>
       </div>
     </Card>
   );